+++ lib/vector_number.rb
@@ -3,8 +3,9 @@
 require_relative "vector_number/version"
 require_relative "vector_number/initializing"
 require_relative "vector_number/mathing"
-require_relative "vector_number/enumerating"
+require_relative "vector_number/math_converting"
 require_relative "vector_number/converting"
+require_relative "vector_number/enumerating"
 require_relative "vector_number/comparing"
 require_relative "vector_number/querying"
 require_relative "vector_number/stringifying"
@@ -13,8 +14,9 @@ require_relative "vector_number/stringifying"
 class VectorNumber
   include Initializing
   include Mathing
-  include Enumerating
+  include MathConverting
   include Converting
+  include Enumerating
   include Comparing
   include Querying
   include Stringifying