+++ spec/spec_helper.rb
@@ -2,6 +2,8 @@
 
 require "vector_number"
 
+require_relative "support/shared_examples"
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"